Rust API: Fix all cargo clippy warnings
Applied automatic fixes and manual corrections for: unnecessary casts
left over from f64->f32 migration, doc comment overindentation,
map_or -> is_none_or/is_some_and, derivable Default impl for
FrameFilter, while_let_loop, iter().copied().collect() -> to_vec(),
and suppressed needless_range_loop where refactoring would reduce
readability.

+++ lib/rust/moteus-protocol/src/multiplex.rs
@@ -186,7 +186,7 @@ impl<'a> WriteCanData<'a> {
                 self.write_i32(saturate_i32(value, scaling.int32));
             }
             Resolution::Float => {
-                self.write_f32(value as f32);
+                self.write_f32(value);
             }
             Resolution::Ignore => {}
         }
